Refactor renderer.
